requirements.txt
streamlit
sentence-transformers
faiss-cpu
pypdf
python-docx
markdown
numpy
ollama
openai
cryptography
python-dotenv
requests

